Support STOR command for uploading files to server
We also removed the explicit /tmp/ directory from the TempDir command to
allow the server to use an OS-specific temporary directory. See:

https://golang.org/pkg/io/ioutil/#TempDir

Listing file directories are a bit more robust. We now clean the
argument sent and only list a directory if indeed that directory exists.
Otherwise, if it is a file, we list only that file and move on. We no
longer list the parent directory unless no argument is sent.

We added the same path cleanup and absolute function call to arguments
that are used in arguments to avoid possibly invalid arguments or
locations that should not be accessed.

Lastly, we now support the STOR command which uses a similar pattern
like retrieving a message.

All thats left is some added error handling, logging, and perhaps some
additional testing. But this is the feature set orignally planned.